Move Mathjax asset to lib directory
