+++ apps/prairielearn/assets/scripts/lib/mathjax.js
